separate out customer details breadcrumb resolver
