+++ apps/breadcrumbs-example/src/app/app.routes.ts
@@ -1,10 +1,8 @@
-import { inject } from '@angular/core';
-import { ActivatedRouteSnapshot, Route } from '@angular/router';
+import { Route } from '@angular/router';
 import { breadcrumbRoute } from '@infinum/ngx-nuts-and-bolts/breadcrumbs';
-import { map } from 'rxjs';
+import { customerDetailsBreadcrumbResolver } from './breadcrumb-resolvers/customer-details.breadcrumb-resolver';
 import { HomepageComponent } from './pages/homepage/homepage.component';
 import { CUSTOMERS_ROUTE_PATH, CUSTOMER_ID_ROUTE_PARAM } from './route-param';
-import { CustomersService } from './services/customers.service';
 
 export const appRoutes: Array<Route> = [
 	breadcrumbRoute({
@@ -23,16 +21,7 @@ export const appRoutes: Array<Route> = [
 			}),
 			breadcrumbRoute({
 				path: `:${CUSTOMER_ID_ROUTE_PARAM}`,
-				breadcrumbResolver: (route: ActivatedRouteSnapshot) => {
-					const customersService = inject(CustomersService);
-
-					const customerId = route.paramMap.get(CUSTOMER_ID_ROUTE_PARAM);
-					if (!customerId) {
-						return 'Customer not found';
-					}
-
-					return customersService.getCustomerById(customerId).pipe(map((customer) => customer.name));
-				},
+				breadcrumbResolver: customerDetailsBreadcrumbResolver,
 				loadComponent: () =>
 					import('./pages/customer-details/customer-details.component').then((m) => m.CustomerDetailsComponent),
 			}),
